moved update functions to service
