+++ ZombieService/NLog.config
@@ -1,0 +1,24 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<nlog xmlns="http://www.nlog-project.org/schemas/NLog.xsd"
+      xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+      autoReload="true"
+      throwExceptions="false">
+
+  <variable name="logDirectory" value="${basedir}/logs"/>
+
+  <targets async="true">
+    <target xsi:type="File"
+            name="default"
+            layout="${longdate}|${level:uppercase=true}|${logger}|${message}"
+            fileName="${logDirectory}\Debug.log"
+            keepFileOpen="false"
+            archiveFileName="${logDirectory}\Debug_${shortdate}.{##}.log"
+            archiveNumbering="Sequence"
+            archiveEvery="Day"
+            maxArchiveFiles="30"
+    />
+  </targets>
+  <rules>
+    <logger name="*" writeTo="default" minlevel="Trace" />
+  </rules>
+</nlog>
