+++ ZombieUtilities/FilePathUtils.cs
@@ -1,4 +1,9 @@
 ﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Management;
+using System.Security.Principal;
+using Microsoft.Win32;
 
 namespace Zombie.Utilities
 {
@@ -27,7 +32,14 @@ namespace Zombie.Utilities
         /// <returns></returns>
         public static string CreateUserSpecificPath(string filePath)
         {
-            var userPath = Environment.GetFolderPath(Environment.SpecialFolder.UserProfile);
+            // (Konrad) Since WindowsService runs in a System context, most of the usual info like Environment.UserName
+            // is not available. This method will return a user name in a Network\username format so we need to parse it.
+            var searcher = new ManagementObjectSearcher("SELECT UserName FROM Win32_ComputerSystem");
+            var networkUsername = (string)searcher.Get().Cast<ManagementBaseObject>().First()["UserName"];
+            var user = networkUsername.Substring(networkUsername.LastIndexOf('\\') + 1);
+
+            var userPath = @"C:\Users\" + user;
+            //var userPath = Environment.GetFolderPath(Environment.SpecialFolder.UserProfile);
             if (filePath.StartsWith("%userpath%"))
             {
                 filePath = filePath.Replace("%userpath%", userPath);
